+++ src/middleware/index.js
@@ -4,12 +4,11 @@
  * @example import { api } from 'middleware';
  */
 
-
 export api from '@plone/volto/middleware/api';
 export crashReporter from '@plone/volto/middleware/crashReporter';
 export blacklistRoutes from './blacklistRoutes';
 export {
   protectLoadStart,
   protectLoadEnd,
-  loadProtector
-} from '@plone/volto/middleware/storeProtectLoadUtils'
+  loadProtector,
+} from '@plone/volto/middleware/storeProtectLoadUtils';